Complete Note Feature. Add phone mask in student feature, add string resources
Need to add snackbar, shimmer loading, change color scheme, integrate homework status, profile fragment, add margins and paddings in layouts, also update recyclers with ItemAnimator, SwipeToDelete etc.

+++ features/students/src/main/kotlin/dev/hawk0f/itutor/features/students/presentation/ui/viewmodels/EditStudentViewModel.kt
@@ -37,8 +37,8 @@ class EditStudentViewModel @Inject constructor(private val getStudentByIdUseCase
         surname = student.fullName.split(' ')[1]
         age = student.age.toString()
         phoneNumber = student.phoneNumber
-        singlePrice = student.singlePrice.toString()
-        groupPrice = student.groupPrice.toString()
+        singlePrice = student.singlePrice.toInt().toString()
+        groupPrice = student.groupPrice.toInt().toString()
         note = student.note
         userId = student.userId
     }
